mkdocs: Disable fetching repository information via GitHub API
As I do not create Git tags in this repository, I do not create releases in
GitHub as well. Material for MkDocs tries to fetch the tag of the latest release
from GitHub, but as it does not exist, fails with status code 404.

Therefore, disable this functionality for now.

See: https://squidfunk.github.io/mkdocs-material/setup/adding-a-git-repository/#adding-a-git-repository.